A creative twist on the classic game Snake
A game of Snake with a colorful twist. In order for the snake to eat the
food, the screen color must match the food. Eat the same color food in a
row and earn more points. Use WASD to control the snake. Use JKL to
change the screen color.

Currently options and some other features are a WIP, but the base game
itself is done.